upnpheader to header mod
